src/requirements.txt
django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-cors-headers==3.12.0
django-utils-six==2.0
coreapi==2.3.3
gunicorn==20.1.0
psycopg2-binary==2.9.3
tzdata==2022.1
Celery[redis]
openpyxl==3.0.10